CHROMIUM: dma-mapping: Add function for dma_addr_t->virt address in device coherent pool
The dma address from device coherent pool can't be translated back to
virtual address easily from outside of kernel/dma/coherent.c, since the
struct dma_coherent_mem and related functions are not exported.
But as in CL:1861475, in videobuf2, we would call dma_sync_* functions
on the dma address obtained from device coherent pool, which need to
know the virtual address from the dma address to flush the cache.

Add a function to translate dma address to virtual address if the
address is from device coherent pool.

BUG=b:141592170
TEST=build

Change-Id: I5d594f128d89168f5851a275595f8f97ac47d7a1
Signed-off-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1861476
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
[rebase510(groeck):
	Function declarations moved from include/linux/dma-mapping.h
	to include/linux/dma-map-ops.h
]
Signed-off-by: Guenter Roeck <groeck@chromium.org>