CHROMIUM: Revert "kbuild: Fail if gold linker is detected"
This reverts commit 75959d44f9dc8e44410667009724e4e238515502.

While the kernel indeed fails to build with the 'gold' linker,
scripts such as kernelconfig use the system default linker,
and that still happens to be gold in Google's development
environment. Revert this patch instead of trying to work around
the problem by specifying per-architecture non-gold linkers
(which would be a pain to implement and to maintain).

BUG=chromium:1006306
TEST=Run "chromeos/scripts/kernelconfig olddefconfig"

Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 6adebd1fb85a592931078f13ca02df814f1698b3)
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Conflicts:
	scripts/Kconfig.include
[rebase510(groeck): Context conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>

Change-Id: I0f274a484164f8261012635edc12e13bb6637b6b