ANDROID: sound: usb: Ensure UAC3 Power Domain is set to D0 when enabling stream
Since commit a0a4959eb4e9 ("ALSA: usb-audio: Operate UAC3 Power
Domains in PCM callbacks") the UAC3 PD state is brought to D1 when
resuming from suspend, and suggests that only when a stream is
begun should a device be put into D0. We need to ensure that is
also done for the tunnel mode case, so add it to
snd_usb_enable_audio_stream() which will get called when receiving
a QMI request to enable the stream.

Fixes: 3f6054629309 ("ANDROID: sound: usb: Add helper APIs to enable audio stream")
Change-Id: I9d3aaf6083a6d859e7e64d167e3ee7eb61522052
Signed-off-by: Jack Pham <jackp@codeaurora.org>
Signed-off-by: Matthias Maennich <maennich@google.com>