FROMLIST: ACPI: Add a convenience function to tell a device is in low power state
Add a convenience function to tell whether a device is in low power state,
primarily for use in drivers' probe or remove functions on busses where
the custom is to power on the device for the duration of both.

Returns false on non-ACPI systems.

Suggested-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(am from https://patchwork.kernel.org/patch/11752757/)

BUG=b:147689556
TEST=Sanity checked basic camera functions

Change-Id: I4694d862a6db5395ec4a691224e17afab97503de
Signed-off-by: Bingbu Cao <bingbu.cao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2403382
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Commit-Queue: Sergey Senozhatsky <senozhatsky@chromium.org>
Tested-by: Sergey Senozhatsky <senozhatsky@chromium.org>