CHROMIUM: config: Turn on CONFIG_IRQ_TIME_ACCOUNTING for arm64
Upstream the arm64 defconfig has this on.  Turning it on appears to
help workloads are mostly IRQ bound, like reading from eMMC while the
device is mostly idle.  A simple eMMC read test on one machine showed
a performance improvement from 225.8 MB/s to 247.1 MB/s.  Presumably
this also may help out some WiFi tests and or graphics tests.

This sounds as if it could cost a little bit of power, but presumably
the performance gain is worth it.

BUG=b:162365396
TEST=eMMC read tests; keep an eye on power_LoadTest.

Change-Id: If56a72059a7bfc7a5b5de8435d77de9e4444470a
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2412648
Reviewed-by: Joel Fernandes <joelaf@google.com>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>