FROMGIT: remoteproc: qcom_q6v5_mss: Allow replacing regulators with power domains
Newer platforms vote for necessary power domains through the power
domain subsystem. For historical reasons older platforms like MSM8916
or MSM8974 still control these as regulators.

Managing them as power domains is preferred since that allows us
to vote for corners instead of raw voltages.

Make it possible for MSM8916 and MSM8974 to manage these as power
domains. For compatibility with old device trees we still need to
support falling back to the regulators when necessary.

The way this is implemented here is that the deprecated regulators
are defined as "fallback_proxy_supply". Only if attaching the power
domains fails because they are not specified (-ENODATA) we request
and manage the fallback regulators instead.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
Link: https://lore.kernel.org/r/20200916104135.25085-7-stephan@gerhold.net
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit 8750cf39239404abbdf727f246723d4e03b54bab
 https://git.kernel.org/pub/scm/linux/kernel/git/andersson/remoteproc.git/ rproc-next)

BUG=b:149924251
TEST=Boot trogdor modem

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: I0579b5e117180dc6b56a4f3e43ade8eefc5838ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2566433
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Sujit Kautkar <sujitka@chromium.org>