CHROMIUM: ASoC: qcom: sc7180: Add PM ops
Add missing PM ops.
This patch should be in v13 if v12 is accepted, or be a FIXUP for v12.

Issue happens when suspend/resume bypasses powerd.
Audio server is not aware of suspend/resume.
After resume, stream will be in RUNNING state, but hw_ptr will not be
moving, thus, no LRCLK. This will cause over-current when SD_MODE is on.

This patch provides the protection to set SD_MODE is off.

Add the PM ops for machine driver so PCM state is set to SUSPENDED.
Note that the state after resume is still not correct, but it can turn
SD_MODE to off.

The root cause and proper solution are under investigation.
Still, machine driver should have PM ops like other machine drivers.

BUG=b:171773926
TEST=Check PCM status when suspend/resume bypasses powerd
 1. aplay -f dat /dev/zero
 2. echo mem > /sys/power/state
 3. press keyboard to resume
 4. check PCM state is (null) instead of RUNNING.
 5. check SD_MODE is 0.

Change-Id: I97c9d0891761282b850697cf9e79e1fe85df7043
Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2502528
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>