FROMGIT: net: ipa: eliminate legacy arguments
We enable a channel doorbell engine only for IPA v3.5.1, and that is
now handled directly by gsi_channel_program().

When initially setting up a channel, we want that doorbell engine
enabled, and we can request that independent of the IPA version.

Doing that makes the "legacy" argument to gsi_channel_setup_one()
unnecessary.  And with that gone we can get rid of the "legacy"
argument to gsi_channel_setup(), and gsi_setup() as well.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit d387c761fa8c9026cd87b8a56a3150535205be73
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git master)

BUG=b:174161909
TEST=Booted trogdor, modem call functional, s/r work

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: Icd9ac645470bf8d59001b141c48d23cc23ad7e7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2558866
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>